refactoring to use TreeNodeVisitor
