+++ crates/integrations/datafusion/src/physical_plan/mod.rs
@@ -15,5 +15,5 @@
 // specific language governing permissions and limitations
 // under the License.
 
-pub(crate) mod predicate_converter;
+pub(crate) mod expr_to_predicate;
 pub(crate) mod scan;
